Django-Backend-Application/requirements.txt
Django==1.11
djangorestframework==3.6.2
pytz==2018.5
